agents/workload_vision/metric_analyzer.py
import cv2
import numpy as np
# import pytesseract
from typing import Dict, Any, Tuple
import logging

logger = logging.getLogger(__name__)

class MetricAnalyzer:
    """
    OpenCV tool to parse and segment screenshots extracting charts and text values.
    """
    def __init__(self):
        pass

    def preprocess_image(self, path: str) -> np.ndarray:
        """Reads, grayscales, and normalizes an image."""
        img = cv2.imread(path, cv2.IMREAD_GRAYSCALE)
        if img is None:
            raise FileNotFoundError(f"Screenshot {path} not found.")
        
        img = cv2.resize(img, (256, 256))
        return img / 255.0

    def extract_graph_regions(self, img: np.ndarray) -> np.ndarray:
        """
        Uses Hough Transforms/Contour detection to isolate line graphs from the UI.
        Placeholder implementation isolating the bottom 70% of the image.
        """
        height, width = img.shape
        cropped = img[int(height * 0.3):height, :]
        return cropped

    def parse_text_metrics(self, path: str) -> Dict[str, str]:
        """
        Uses pytesseract to OCR specific bounding boxes where KPIs are known to be.
        """
        logger.info("OCR Parsing image variables...")
        # Placeholder for Tesseract
        # text = pytesseract.image_to_string(img)
        return {"detected_cpu": "89%", "detected_latency": "1402ms"}
